# [TESTERS ONLY][Experimental] rsx: Experimental "Fast" RSX mode