Make the CELL think RSX already finished its work so it can already prepare the next drawing state, meanwhile, RSX executes the previous draws submitted by CELL and therefore no side is stalling the other (at least that's the goal). This only applies when using Atomic FIFO accuracy atm. In theory, this may help to utilize more threads for high-threaded CPUs.
Test performance and CPU usage before and after with different games.