@solarmystic 2022-08-24
**No significant changes in performance noted in games below. Tested on a 12400F/2080TI rig.**

**Persona 5 (1F Classroom Building Hallway)**

PR - 78/63 1%/57 0.1% FPS - 59% CPU Usage
![P5 PR](https://user-images.githubusercontent.com/4345150/186286696-d993419a-1877-42a2-9f8c-dd8f514a4ab5.PNG)

Current master (14084) - 79/64 1%/58 0.1 FPS - 64% CPU Usage
![P5 14084](https://user-images.githubusercontent.com/4345150/186286709-20cdbfcc-0757-43f5-80c4-8f142559f0d4.PNG)


**God of War 3 (Three Judges Savepoint)** 

PR - 100/74 1%/58 0.1% FPS - 51% CPU Usage
![GOW3 PR](https://user-images.githubusercontent.com/4345150/186286728-d572d67d-1949-4ac1-baef-461b61e3aa32.PNG)

Current master (14084) - 100/74 1%/59 0.1 FPS - 51% CPU Usage
![GOW3 14084](https://user-images.githubusercontent.com/4345150/186286732-dfd31458-e451-430f-a12c-d079fac98658.PNG)


**God of War Ascension (chapter 1 start)**

PR - 45/40 1%/38 0.1% FPS - 63% CPU Usage
![GOWA PR](https://user-images.githubusercontent.com/4345150/186286764-4ae5ac2c-f418-422a-a977-8c9381b95f18.PNG)

Current master (14084) - 45/39 1%/37 0.1% FPS - 63% CPU Usage
![GOWA 14084](https://user-images.githubusercontent.com/4345150/186286779-66619803-4c14-4ceb-9a63-75adfd593305.PNG)


**The Last Of Us (The Woods)**

PR - 36/32 1%/31 0.1 FPS - 62% CPU Usage
![TLOU PR](https://user-images.githubusercontent.com/4345150/186286791-8943e3ae-fa96-42b0-a964-8a5f757b2f02.PNG)

Current master (14084) - 36/32 1%/31 0.1 FPS - 62% CPU Usage
![TLOU 14084](https://user-images.githubusercontent.com/4345150/186286799-ba42dc65-df6b-4738-b07b-cf405e9032b2.PNG)


**Demon's Souls**

PR - 85/62 1%/58 0.1% FPS - 34% CPU Usage
![DS PR](https://user-images.githubusercontent.com/4345150/186286823-ebfcabf9-9052-4078-a5d2-5358f639efad.PNG)

Current master (14084) - 85/62 1%/59 0.1% FPS - 35% CPU Usage
![DS 14084](https://user-images.githubusercontent.com/4345150/186286831-78640225-e894-41ec-ae35-bb8bc91028ef.PNG)